Tighten up the spacing between sidebar sections
Set sidebar sections and subtitles to be a consistent amount of
margin away from each other. Also set this so that it adjusts
based on the screen size (aka: on mobile we don’t want this to
result in unnecessary scrolling). Note the subtitles are now
closer to the content and sections are now farther away from
each other. A share-link class was added to shrink the margin
between the share/join link description and the link itself. Move
the description closer to the side bar title.