Turn share/invite group into a sidebar section
Put the share/invite info into a sidebar section to standardize
the styling and re-use the subtitle css class rather than
using/creating a separate class that effectively does the
same thing. Note this change deletes the old css title class
as it’s no longer needed. In order to maintain the same
margin between the subtitle of the section and the paragraph
tag, some margin and padding was added on top to the share-link
css class.